COMP: Build python with zlib support
This will allow the extraction of "*.tar.gz" file using the "tarfile" python
module.

On unix system, considering the API/ABI of zlib library is stable,
we can assume the zlib library provided by the system will be sufficient.

On windows, there are no zlib library available on the system. That is why
it is superbuild and build python against it.

See http://na-mic.org/Bug/view.php?id=3055
Fixes #3055