COMP: Properly fixed up python extensions
Before this commit python extensions were not properly fixed up. For
example, within a MacOSX package, running "otool -L" on any of the
python extensions located in "lib-dynload" folder was outputting the
following:

$ otool -L _tkinter.so
_tkinter.so:
    @executable_path/../../Contents/MacOS/_tkinter.so (compatibility version 0.0.0, current version 0.0.0)
    @executable_path/../lib/TclTk/lib/libtcl8.4.dylib (compatibility version 8.4.0, current version 8.4.19)
    @executable_path/../lib/TclTk/lib/libtk8.4.dylib (compatibility version 8.4.0, current version 8.4.19)
    /usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 125.2.0)

$ otool -L _csv.so
_csv.so:
    @executable_path/../../Contents/MacOS/_csv.so (compatibility version 0.0.0, current version 0.0.0)
    /usr/lib/libSystem.B.dylib (compatibility version 1.0.0, current version 125.2.0)


whereas the expected output is:

  @executable_path/../../Contents/lib/Python/lib/python2.7/lib-dynload/<PythonExtension>.so


Since the extensions being fixed up are copied from the python-install
folder, we now make sure the path are not stripped by setting CMAKE_INSTALL_TOOL
to an empty string.