ENH: RSS module - Add rejection label
Before this commit, the RSS module works as: given a seed label image, the surface grows from the seeded region and grab the target region.
But in case of leakage, it is not able to prevent such leakage by assigning some rejection region.

This commit introduced the possibility of rejection label. There is no added parameter. Just in the input label image, use another color(label) to mark the rejection region.

say we draw a seed label image using two colors/labels: 1 and 2, 10, 24. Then the contour will grow from label 1 (label of interest, default is 1 and can be adjusted)
and moreover it will NOT grow into label 2/10/24.