STYLE: Remove unused Initialize methods from vtkMRMLSliceLogic API
The methods "Initialize" and "IsInitialized" have been first introduced
in r14218 as a general way to initialize logic object and then
their usage has been depreacted and removed by commit r15928 and r15929.

It turns out that in this case, a two step initialization was not needed
and the same goal could be achieved using methods with explicit name.
In that specific case, methods "setMRMLSliceNodeInternal"
and "setMRMLSliceCompositeNodeInternal".