COMP: Move VTKv6 and ITKv4 find_package calls into each Slicer sub-project.
To address issue #3741 [1], call to find_package associated to VTK and ITK
have been moved in every Slicer sub project, library or module expecting it.

This allows to reduce both the number of include directories and
the number of libraries passed to the compiler.

Moving forward, every Slicer sub project, library or module will be
responsible to call find_package (optionally specifying a list of
components) and (if it applies) include the associated "use file".

[1] http://na-mic.org/Mantis/view.php?id=3741