COMP: Install VTKv6 and ITKv4 projects to reduce compiler/linker input.
With introduction of modularization in ITK and VTK, the number of include
directories increased a lot and this became an issue to build Slicer
extension.

This commit is an attempt to fix build error related to SlicerIGT extension:

  execvp: /bin/sh: Argument list too long

Finally, let's note that moving forward slicer modules and libraries
will be expected to call the find_package for their explicit dependency.
See commit r23353 and [1] for more details.

[1] http://massmail.spl.harvard.edu/public-archives/slicer-devel/2014/015671.html

Conflicts:
	SuperBuild/External_ITKv4.cmake
	SuperBuild/External_VTKv6.cmake