STYLE: Remove deprecated variable from CDash driver script
Setting ADDITIONAL_{C|CXX}_FLAGS variables is not the conventional
way of setting additional flags. Setting environment variable
CFLAGS and CXXFLAGS is the recommended way.

Remove variable "WITH_TESTING_EXTENSIONS". This is an advanced variable
that shouldn't be enabled on contributed dashboards. Only the factories
should eventually enable it. That said, the recommended way is to
build a set of s4ext files as explained in [1]

[1] http://wiki.slicer.org/slicerWiki/index.php/Documentation/Nightly/Developers/Build_ExtensionsIndex