BUG: Fix tests by addressing regression introduced in r23788. See #3678
* Introduce a convenience static function vtkMRMLSelectionNode::GetSelectionNode
allowing to get the singleton selection node associated with a scene if any.

* Remove caching of selection node in qSlicerModelsModuleWidget,
qMRMLModelDisplayNodeWidget and vtkMRMLModelDisplayableManager. For sake
of simplicity, the selection node is systematically retrieve using the method
vtkMRMLSelectionNode::GetSelectionNode.

* Properly QString::fromStdString by assigning the function output the
class member "FiberDisplayClass" instead of simply calling function on
the member.

This commit fixes the following tests (tested on Ubuntu 13.10 / Debug
build):

* qSlicerModelsModuleWidgetGenericTest
* qSlicerModelsModuleGenericTest
* qSlicerModelsModuleWidgetTest
* qSlicerModelsModuleWidgetTest1
* py_AtlasTests
* py_AbdominalAtlasTest
* py_KneeAtlasTest
* py_ScenePerformance