STYLE: vtkMRMLSliceNode: Remove unused SetOrientationString
The method is not used in any Slicer extensions.

Notes:

(1) explicitly setting the orienation string to "Reformat" is
not not needed anymore because "GetOrientation()" or "GetOrientationString()"
will return "Reformat" if the SliceToRAS matrix is not one of the preset.

(2) For sake of consistency, in the RenameSliceOrientationPreset, the
orientation reference is also renamed.

(3) In "Copy()" method, Orientation and OrientationReference are set
after the presets.