ENH: Update to setuptools v35.0.1
This commit updates to the latest version of setuptools and removes the now
unneeded patch for distutils.msvc9compiler. Indeed, latest version of
setuptools now properly patch distutils so that it works well with most
version of visual studio.

This commit also explicitly introduces the direct and indirect setuptools
dependencies:
* python-appdirs
* python-packaging
* python-pyparsing
* python-six

Note that setuptools is first bootstrapped, this is required to ensure
setuptools is installed properly with the file setuptools-X.Y.Z-py2.7.egg/EGG-INFO/entry_points.txt
available.