ENH: Integrate VTK plotting to support Interactive manipulation
This commit provides Plotting facilities that include a Plot View
that can be packed in the layout, similar to the Slice Views and 3D Views.

The architecture also includes nodes that represent the data to be displayed
in the plot as well as nodes to represent the display properties of the data
and the plot itself.

In Addition, the PlotView is connected with the Table Module and Views allowing
the plotting of arrays with a simple click.

References:
https://www.slicer.org/wiki/Documentation/Nightly/Developers/Plots
https://www.na-mic.org/wiki/Project_Week_25/Interactive_Manipulation_of_Plots_and_Graphs
https://www.slicer.org/wiki/Developer_Meetings/20170425