BUG: Fix initialization of module displayable managers
This commit removes the auto-initialization of displayable managers when
the displayable manager library was loaded. This approach was failing
when Slicer was built without Python support. Indeed, in that case the
displayable manager libraries were not loaded as a side effect of importing
the wrapper libraries.

Instead, this commit requires module to explicitly initialize its
associated displayable managers (similarly to what is already done in
tests.